requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.29.0
SQLAlchemy>=2.0.29,<2.1
pydantic>=2.6.4,<3.0
python-dotenv>=1.0.1,<2.0
cryptography>=41.0.7,<43.0
PyJWT>=2.8.0,<3.0
typer>=0.9.0,<0.11.0
